REFONTE-BACK Add JWT

+++ src/main/java/com/circe/invoice/repository/data/CompanyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompanyRepository extends JpaRepository<CompanyEntity, Long> {
+public interface CompanyRepository extends JpaRepository<CompanyEntity, Integer> {
 
     CompanyEntity findByName(String name);
 }
