+++ src/main/java/com/circe/invoice/repository/referential/AuthorityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthorityRepository extends JpaRepository<AuthorityEntity, Long> {
+public interface AuthorityRepository extends JpaRepository<AuthorityEntity, Integer> {
     AuthorityEntity findByName(String name);
 }
